[IMPLEMENT] Implement maker renew (main feature).

+++ Controller/MakerController.php
@@ -1,35 +1,40 @@
 <?php
 /*
-* This file is part of EC-CUBE
-*
-* Copyright(c) 2000-2015 LOCKON CO.,LTD. All Rights Reserved.
-* http://www.lockon.co.jp/
-*
-* For the full copyright and license information, please view the LICENSE
-* file that was distributed with this source code.
-*/
+ * This file is part of the Maker plugin
+ *
+ * Copyright (C) 2016 LOCKON CO.,LTD. All Rights Reserved.
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
 
 namespace Plugin\Maker\Controller;
 
-use Plugin\Maker\Form\Type\MakerType;
+use Doctrine\Common\Collections\ArrayCollection;
 use Eccube\Application;
+use Plugin\Maker\Entity\Maker;
 use Symfony\Component\HttpFoundation\Request;
-use Symfony\Component\HttpKernel\Exception as HttpException;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
+/**
+ * Class MakerController
+ * @package Plugin\Maker\Controller
+ */
 class MakerController
 {
-    private $main_title;
-    private $sub_title;
-
-    public function __construct()
+    /**
+     * List, add, edit maker
+     *
+     * @param Application $app
+     * @param Request     $request
+     * @param null        $id
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse|\Symfony\Component\HttpFoundation\Response
+     */
+    public function index(Application $app, Request $request, $id = null)
     {
-    }
+        $repos = $app['eccube.plugin.maker.repository.maker'];
 
-    public function index(Application $app, Request $request, $id)
-    {
-    	$repos = $app['eccube.plugin.maker.repository.maker'];
-
-		$TargetMaker = new \Plugin\Maker\Entity\Maker();
+        $TargetMaker = new Maker();
 
         if ($id) {
             $TargetMaker = $repos->find($id);
@@ -42,36 +47,45 @@ class MakerController
             ->createBuilder('admin_maker', $TargetMaker)
             ->getForm();
 
-        if ('POST' === $request->getMethod()) {
-            $form->handleRequest($request);
+        $form->handleRequest($request);
 
-            if ($form->isValid()) {
-                $status = $repos->save($TargetMaker);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $status = $repos->save($TargetMaker);
 
-                if ($status) {
-                    $app->addSuccess('admin.maker.save.complete', 'admin');
-                    return $app->redirect($app->url('admin_maker'));
-                } else {
-                    $app->addError('admin.maker.save.error', 'admin');
-                }
+            if ($status) {
+                $app->addSuccess('admin.maker.save.complete', 'admin');
+                return $app->redirect($app->url('admin_maker'));
+            } else {
+                $app->addError('admin.maker.save.error', 'admin');
             }
         }
-    	
-        $Makers = $app['eccube.plugin.maker.repository.maker']->findAll();
 
-        return $app->render('Maker/View/admin/maker.twig', array(
+        /**
+         * @var ArrayCollection
+         */
+        $arrMaker = $app['eccube.plugin.maker.repository.maker']->findBy(array(), array('rank' => 'DESC'));
+
+        return $app->render('Maker/Resource/template/admin/maker.twig', array(
         	'form'   		=> $form->createView(),
-            'Makers' 		=> $Makers,
+            'arrMaker' 		=> $arrMaker,
             'TargetMaker' 	=> $TargetMaker,
         ));
     }
 
+    /**
+     * Delete Maker
+     *
+     * @param Application $app
+     * @param Request     $request
+     * @param integer     $id
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     */
     public function delete(Application $app, Request $request, $id)
     {
-    	$repos = $app['eccube.plugin.maker.repository.maker'];
+        $repos = $app['eccube.plugin.maker.repository.maker'];
 
         $TargetMaker = $repos->find($id);
-        
+
         if (!$TargetMaker) {
             throw new NotFoundHttpException();
         }
@@ -83,11 +97,9 @@ class MakerController
             ->getForm();
 
         $status = false;
-        if ('POST' === $request->getMethod()) {
-            $form->handleRequest($request);
-            if ($form->isValid()) {
-                $status = $repos->delete($TargetMaker);
-            }
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $status = $repos->delete($TargetMaker);
         }
 
         if ($status === true) {
@@ -99,10 +111,18 @@ class MakerController
         return $app->redirect($app->url('admin_maker'));
     }
 
+    /**
+     * Up rank
+     *
+     * @param Application $app
+     * @param Request     $request
+     * @param integer     $id
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     */
     public function up(Application $app, Request $request, $id)
     {
-    	$repos = $app['eccube.plugin.maker.repository.maker'];
-    	
+        $repos = $app['eccube.plugin.maker.repository.maker'];
+
         $TargetMaker = $repos->find($id);
         if (!$TargetMaker) {
             throw new NotFoundHttpException();
@@ -115,11 +135,9 @@ class MakerController
             ->getForm();
 
         $status = false;
-        if ('POST' === $request->getMethod()) {
-            $form->handleRequest($request);
-            if ($form->isValid()) {
-                $status = $repos->up($TargetMaker);
-            }
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $status = $repos->up($TargetMaker);
         }
 
         if ($status === true) {
@@ -131,10 +149,18 @@ class MakerController
         return $app->redirect($app->url('admin_maker'));
     }
 
+    /**
+     * Down rank
+     *
+     * @param Application $app
+     * @param Request     $request
+     * @param integer     $id
+     * @return \Symfony\Component\HttpFoundation\RedirectResponse
+     */
     public function down(Application $app, Request $request, $id)
     {
-    	$repos = $app['eccube.plugin.maker.repository.maker'];
-    	
+        $repos = $app['eccube.plugin.maker.repository.maker'];
+
         $TargetMaker = $repos->find($id);
         if (!$TargetMaker) {
             throw new NotFoundHttpException();
@@ -147,11 +173,9 @@ class MakerController
             ->getForm();
 
         $status = false;
-        if ('POST' === $request->getMethod()) {
-            $form->handleRequest($request);
-            if ($form->isValid()) {
-                $status = $repos->down($TargetMaker);
-            }
+        $form->handleRequest($request);
+        if ($form->isSubmitted() && $form->isValid()) {
+            $status = $repos->down($TargetMaker);
         }
 
         if ($status === true) {
@@ -162,5 +186,4 @@ class MakerController
 
         return $app->redirect($app->url('admin_maker'));
     }
-
 }
